requirements.txt
python-dotenv
openai
google-generativeai
requests
pandas
openpyxl